requirements.txt
numpy>=1.19.0
scipy>=1.5.0
matplotlib>=3.3.0
Pillow>=8.0.0